backend/requirements.txt
# FastAPI
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2

# ML & Image Processing
torch==2.1.2
torchvision==0.16.2
pillow==10.2.0
numpy==1.26.3
scikit-learn==1.3.2

# External APIs
google-generativeai==0.3.2
httpx==0.26.0

# Storage
cloudinary==1.38.0
# boto3==1.34.34  # Uncomment for AWS S3

# Utilities
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0
